+++ go_router/lib/src/go_router_delegate.dart
@@ -159,21 +159,21 @@ class GoRouterDelegate extends RouterDelegate<Uri>
   void go(String location, {Object? extra}) {
     log.info('going to $location');
     _go(location, extra: extra);
-    _notifyListeners();
+    notifyListeners();
   }
 
   /// push the given location onto the page stack
   void push(String location, {Object? extra}) {
     log.info('pushing $location');
     _push(location, extra: extra);
-    _notifyListeners();
+    notifyListeners();
   }
 
   /// Refresh the current location, including re-evaluating redirections.
   void refresh() {
     log.info('refreshing $location');
     _go(location, extra: _matches.last.extra);
-    _notifyListeners();
+    notifyListeners();
   }
 
   /// Get the current location, e.g. /family/f2/person/p1
@@ -347,7 +347,7 @@ class GoRouterDelegate extends RouterDelegate<Uri>
 
         // let Router know to update the address bar
         // (the initial route is not a redirect)
-        if (redirects.length > 1) _notifyListeners();
+        if (redirects.length > 1) notifyListeners();
 
         // no more redirects!
         break;
@@ -585,14 +585,15 @@ class GoRouterDelegate extends RouterDelegate<Uri>
   }
 
   Widget _builder(BuildContext context, Iterable<GoRouteMatch> matches) {
-    var pages = <Page<dynamic>>[];
+    List<Page<dynamic>>? pages;
 
     try {
       // build the stack of pages
       if (routerNeglect) {
-        Router.neglect(context, () {
-          pages = getPages(context, matches.toList()).toList();
-        });
+        Router.neglect(
+          context,
+          () => pages = getPages(context, matches.toList()).toList(),
+        );
       } else {
         pages = getPages(context, matches.toList()).toList();
       }
@@ -621,14 +622,16 @@ class GoRouterDelegate extends RouterDelegate<Uri>
       ];
     }
 
+    // we should've set pages to something by now
+    assert(pages != null);
+
     // wrap the returned Navigator to enable GoRouter.of(context).go()
     return builderWithNav(
       context,
       Navigator(
         restorationScopeId: restorationScopeId,
-        key: _key,
-        // needed to enable Android system Back button
-        pages: pages,
+        key: _key, // needed to enable Android system Back button
+        pages: pages!,
         observers: observers,
         onPopPage: (route, dynamic result) {
           if (!route.didPop(result)) return false;
@@ -643,7 +646,7 @@ class GoRouterDelegate extends RouterDelegate<Uri>
 
           // this hack allows the browser's address bar to be updated after a
           // push and pressing the Back button, but it shouldn't be necessary...
-          _notifyListeners();
+          notifyListeners();
 
           return true;
         },
@@ -846,9 +849,5 @@ class GoRouterDelegate extends RouterDelegate<Uri>
     assert(uri.queryParameters.isEmpty);
     return _canonicalUri(
         Uri(path: uri.path, queryParameters: queryParams).toString());
-  }
-
-  void _notifyListeners() {
-    notifyListeners();
   }
 }